Moved TypeHelper to Microsoft.AspNet.Mvc.Common.
