+++ src/Microsoft.AspNet.Mvc.Common/project.json
@@ -8,6 +8,7 @@
         "aspnet50": { },
         "aspnetcore50": {
             "dependencies": {
+		"System.ComponentModel.TypeConverter": "4.0.0-beta-*",
                 "System.Reflection.Extensions": "4.0.0-beta-*",
                 "System.Text.Encoding.Extensions": "4.0.10-beta-*"
             }
